PYIC-7877 Remove json support
