+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/service/AppConfigService.java
@@ -87,12 +87,6 @@ public class AppConfigService extends YamlParametersConfigService {
         return super.getParametersByPrefix(path);
     }
 
-    @Override
-    public Map<String, String> getParametersByPrefixYaml(String path) {
-        reloadParameters();
-        return super.getParametersByPrefixYaml(path);
-    }
-
     private void reloadParameters() {
         var profileId = getEnvironmentVariable(EnvironmentVariable.APP_CONFIG_PROFILE_ID);
         var paramsRaw = appConfigProvider.get(profileId);
