+++ libs/common-services/src/main/java/uk/gov/di/ipv/core/library/service/SsmConfigService.java
@@ -139,11 +139,6 @@ public class SsmConfigService extends ConfigService {
 
     @Override
     protected Map<String, String> getParametersByPrefix(String path) {
-        return ssmProvider.getMultiple(resolvePath(path));
-    }
-
-    @Override
-    protected Map<String, String> getParametersByPrefixYaml(String path) {
         var parameters = ssmProvider.getMultiple(resolvePath(path));
         if (parameters.isEmpty()) {
             throw new ConfigParameterNotFoundException("SSM parameter not found for path: " + path);
