+++ local-running/core.local.params.yaml
@@ -3,7 +3,6 @@
 ---
 core:
   self:
-    configFormat: "yaml"
     componentId: "https://identity.local.account.gov.uk"
     audienceForClients: "https://identity.local.account.gov.uk"
     jwtTtlSeconds: 3600
